// src/building/types.ts
export enum EffectType {
  STOP_MOVE = 'stopMove', // 禁锢
  ICE = 'ice', // 冰冻
  FIRING = 'firing', // 灼烧
  ADD_FIRING = 'addFiring', // 添加灼烧
  BOMB = 'bomb', // 炸弹
  SHIELD = 'shield', // 护盾
  VENOM = 'venom', // 毒液
  RESTORE = 'restore', // 治疗
}